+++ src/Scheduler/BatchableSchedulerInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phlib\JobQueue\Scheduler;
+
+/**
+ * @package Phlib\JobQueue
+ */
+interface BatchableSchedulerInterface extends SchedulerInterface
+{
+    public function retrieveBatch(): array|false;
+
+    public function removeBatch(array $jobId): bool;
+}